Move Server Object logic away from models code
